feat(suite-common): Move export transaction to suite-common
